Move CustomNonce component in /app folder and used ButtonLink for editing custom nonce
