+++ ui/components/app/app-components.scss
@@ -10,6 +10,7 @@
 @import 'confirm-page-container/index';
 @import 'confirm-data/index';
 @import 'confirmation-warning-modal/index';
+@import 'custom-nonce/index';
 @import 'nfts-items/index';
 @import 'nfts-tab/index';
 @import 'nft-details/index';
