+++ ui/components/ui/ui-components.scss
@@ -13,7 +13,6 @@
 @import 'color-indicator/color-indicator';
 @import 'confusable/index';
 @import 'currency-display/index';
-@import 'custom-nonce/index';
 @import 'definition-list/definition-list';
 @import 'dialog/dialog';
 @import 'dropdown/dropdown';
